backend/python_backend/requirements.txt
fastapi
uvicorn
plaid-python
httpx
python-dotenv
pandas
numpy
pydantic
